Command arguments now handled by internal options module
